+++ server/services/store/sqlstore/data_migrations_test.go
@@ -1,0 +1,1 @@
+package sqlstore
